+++ api/instrument.php
@@ -1,9 +1,4 @@
-<?php // Send information about an instrument.
-
-// 422 = Invalid data
-// 428 = Precondition required
-// 404 = Not found
-// 400 = Bad request
+<?php // Send information about an instrument (Assay Module).
 
 /*
 Get moduleSerialNumber, frontPanelId, status, progress, timeRemaining, faultCode, versionNumber
@@ -17,7 +12,7 @@ Codes:
     - 5: FAULT CONDITION
 
     Fault Code:
-    - TBD
+    - TODO: TBD
 
     Progress:
     - % completed (0-100)
@@ -33,15 +28,28 @@ Codes:
 // Firstly, if we have no data, quit.
 if(!$data) {
     // Throw error and stop processing things.
-    echo(json_encode(["Error" => "No data available."]));
+    echo(json_encode(["status" => 400]));
     exit;
 }
 
+// We have data! Clean it and add it to the instruments db if it is good.
+$instrumentData['serial_number'] = $data['moduleSerialNumber'] ?? null;
+
+if($instrumentData['serial_number']) {
+    $instrumentData['module_id'] = $data['frontPanelId'] ?? null;
+    $instrumentData['status'] = $data['status'] ?? null;
+    $instrumentData['progress'] = $data['progress'] ?? null;
+    $instrumentData['time_remaining'] = $data['timeRemaining'] ?? null;
+    $instrumentData['fault_code'] = $data['faultCode'] ?? null;
+    $instrumentData['version_number'] = $data['versionNumber'] ?? null;
+
+    $response = updateInstrument($instrumentData);
+}
 
 
 // Provide the response
-if($response) {
-    echo json_encode(["status" => $response]);
+if(isset($response)) {
+    echo json_encode(["status" => 10]);
 } else {
-    echo json_encode(["status" => "ERROR"]);
+    echo json_encode(["status" => 422]);
 }